Style: 불필요한 코드 삭제 및 Confetti 파일 위치 변경
